Refactor commit lock mechanism from HiveTableOperations
